pyproject.toml
[project]
name = "content-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.9",
    "mcp[cli]>=1.15.0",
    "openai>=1.109.1",
    # "openai-agents>=0.3.3",  # Removed - not compatible with local LLMs
    "python-dotenv>=1.1.1",
    "youtube-transcript-api>=1.2.2",
]
